+++ just-in-time/bundle-choice-handler/bundleChoiceHandler.test.js
@@ -54,6 +54,20 @@ describe('order confirmation handler test', ()=>{
         expect(sayText).toHaveBeenCalledWith(inputMenu[1]);
         expect(promptDigits).toHaveBeenCalledWith(handlerName);
     });
+    it('should not display a menu page if the input is 77 and the next menu doesn\'t exists',()=>{
+        state.vars.multiple_input_menus = true;
+        state.vars.input_menu_loc = 3;
+        state.vars.input_menu_length = 3;
+        state.vars.input_menu = JSON.stringify(inputMenu);
+        bundleChoiceHandler(77);
+        expect(sayText).not.toHaveBeenCalled();
+    });
+    it('should not set state.vars.chosenMaizeBundle bundle if the quantity is not defined ',()=>{
+        state.vars.multiple_input_menus = false;
+        state.vars.chosenMaizeBundle = undefined;
+        bundleChoiceHandler(2);
+        expect(state.vars.chosenMaizeBundle).toBeUndefined();
+    });
 
 
 });